depuracion v8

+++ core/representantetecnico/views/empresa/delete.py
@@ -27,6 +27,8 @@ class EmpresaDeleteView(LoginRequiredMixin, ValidatePermissionRequiredMixin, Del
     def post(self, request, *args, **kwargs):
         data = {}
         try:
+            if ComprasPublicas.objects.filter(empresa_id=self.object.id).exists():
+                raise Exception('No es posible eliminar este registro')
             self.object.delete()
         except Exception as e:
             data['error'] = str(e)
